Update isMalformed
`isMalformed` used a regex to find invalid characters. This however this has bad performance, since the regex will be created each time `isMalformed` is called. In Addition to that string replacements are a very expensive operations, since multiple string instances are created. 

To improve this behavior we iterate through all the characters in the string and check them against the condition using character methods, and comparing single characters. It is enough to find only one invalid character, so we can abort, when we find one. 

There is still room for improvement: Put all the invalid characters, that cannot easily classified using character classes, into a hashset, which has a lookup in O(1).

Overall the method is in O(n),